Getting Data from REST API
Granted it doesn't look pretty I have data coming in from the REST API
to a simple text view. The data is not parsed and is coming in as a
giant string which was to be expected.

I will be implementing the rest of the get other non parameterized
methods we will need.

We will want to refactor later on to remove the code duplication that
will likely be occuring.