Check Internet Connection
Implemented new permission for Access Network State so we could
implement a logical check to ensure there is a network connection prior
to performing an asynchronous task otherwise we will display an error
message that the internet connection is not available and to try again
later.

Moved all of the hard coded strings into the strings.xml resource.